+++ frontend/src/components/AuthenticatedMain.vue
@@ -9,13 +9,11 @@
     </div>
   </div>
 
-  <div v-else>
-    <NavigationBar :me="me" />
-
+  <AppShell v-else :me="me">
     <div class="max-w-7xl mx-auto px-4 py-12 sm:px-6 lg:px-8">
       <router-view />
     </div>
-  </div>
+  </AppShell>
 </template>
 
 <script setup lang="ts">
@@ -23,8 +21,8 @@ import { onMounted, ref } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { UserDto } from '../common/backend';
 import userdata from '../common/userdata';
+import AppShell from './AppShell.vue';
 import FetchError from './FetchError.vue';
-import NavigationBar from './NavigationBar.vue';
 
 const { t } = useI18n({ useScope: 'global' });